Support creating cert chains
Sometimes, certs are signed by untrusted intermediate CAs, and a cert
along with any intermediates needs to be sent during the TLS handshake.
This adds a convenience so you can create a cert chain by using an
array. This is distinct from supporting multiple site certificates,
which can also themselves be served up as chains.